backend/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
faiss-cpu
sentence-transformers
PyMuPDF
python-dotenv
pydantic-settings
numpy
pytest
httpx
